+++ src/TreeTools.h
@@ -1,0 +1,8 @@
+#ifndef RCPP_TreeTools_RCPPEXPORTS_H_GEN_
+#define RCPP_TreeTools_RCPPEXPORTS_H_GEN_
+#include <Rcpp.h>
+using namespace Rcpp;
+extern IntegerMatrix preorder_edges_and_nodes(IntegerVector parent,
+                                              IntegerVector child);
+
+#endif // RCPP_TreeTools_RCPPEXPORTS_H_GEN_